requirements.txt
fastmcp>=2.0.0
mcp>=1.0.0
starlette>=0.37.2
openai>=1.40.0
sympy>=1.12
numpy>=1.26
pillow>=10.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.1
